# internal/apiv1gen/openapi.gen.go
//go:build go1.22

// Package apiv1gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package apiv1gen

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/oapi-codegen/runtime"
	strictnethttp "github.com/oapi-codegen/runtime/strictmiddleware/nethttp"
)

// Defines values for EventRecordLevel.
const (
	Debug EventRecordLevel = "debug"
	Error EventRecordLevel = "error"
	Info  EventRecordLevel = "info"
	Warn  EventRecordLevel = "warn"
)

// APIResponse defines model for APIResponse.
type APIResponse struct {
	Accepted            bool    `json:"accepted"`
	AcceptedEventCount  *int    `json:"accepted_event_count,omitempty"`
	BatchId             *string `json:"batch_id,omitempty"`
	DuplicateEventCount *int    `json:"duplicate_event_count,omitempty"`
	ErrorCode           *string `json:"error_code,omitempty"`
	Message             *string `json:"message,omitempty"`
	Path                *string `json:"path,omitempty"`
	RejectedEventCount  *int    `json:"rejected_event_count,omitempty"`
}

// BatchRequest defines model for BatchRequest.
type BatchRequest struct {
	BatchId string        `json:"batch_id"`
	Events  []EventRecord `json:"events"`
	SentAt  string        `json:"sent_at"`
}

// EventRecord defines model for EventRecord.
type EventRecord struct {
	Component     string                 `json:"component"`
	Event         string                 `json:"event"`
	EventId       string                 `json:"event_id"`
	Fields        map[string]interface{} `json:"fields"`
	InstallId     string                 `json:"install_id"`
	Level         EventRecordLevel       `json:"level"`
	SchemaVersion int                    `json:"schema_version"`
	Timestamp     string                 `json:"timestamp"`
}

// EventRecordLevel defines model for EventRecord.Level.
type EventRecordLevel string

// HealthResponse defines model for HealthResponse.
type HealthResponse struct {
	Ok bool `json:"ok"`
}

// ReadyResponse defines model for ReadyResponse.
type ReadyResponse struct {
	ErrorCode *string `json:"error_code,omitempty"`
	Ok        bool    `json:"ok"`
}

// XManifoldIngestKey defines model for XManifoldIngestKey.
type XManifoldIngestKey = string

// XRequestId defines model for XRequestId.
type XRequestId = string

// PostEventsBatchParams defines parameters for PostEventsBatch.
type PostEventsBatchParams struct {
	XRequestID         *XRequestId         `json:"X-Request-ID,omitempty"`
	XManifoldIngestKey *XManifoldIngestKey `json:"X-Manifold-Ingest-Key,omitempty"`
}

// PostEventsBatchJSONRequestBody defines body for PostEventsBatch for application/json ContentType.
type PostEventsBatchJSONRequestBody = BatchRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Liveness probe
	// (GET /healthz)
	GetHealthz(w http.ResponseWriter, r *http.Request)
	// Readiness probe backed by storage ping
	// (GET /readyz)
	GetReadyz(w http.ResponseWriter, r *http.Request)
	// Ingest a batch of structured events
	// (POST /v1/events/batch)
	PostEventsBatch(w http.ResponseWriter, r *http.Request, params PostEventsBatchParams)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetHealthz operation middleware
func (siw *ServerInterfaceWrapper) GetHealthz(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetHealthz(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetReadyz operation middleware
func (siw *ServerInterfaceWrapper) GetReadyz(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetReadyz(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PostEventsBatch operation middleware
func (siw *ServerInterfaceWrapper) PostEventsBatch(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params PostEventsBatchParams

	headers := r.Header

	// ------------- Optional header parameter "X-Request-ID" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("X-Request-ID")]; found {
		var XRequestID XRequestId
		n := len(valueList)
		if n != 1 {
			siw.ErrorHandlerFunc(w, r, &TooManyValuesForParamError{ParamName: "X-Request-ID", Count: n})
			return
		}

		err = runtime.BindStyledParameterWithOptions("simple", "X-Request-ID", valueList[0], &XRequestID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: false})
		if err != nil {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "X-Request-ID", Err: err})
			return
		}

		params.XRequestID = &XRequestID

	}

	// ------------- Optional header parameter "X-Manifold-Ingest-Key" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("X-Manifold-Ingest-Key")]; found {
		var XManifoldIngestKey XManifoldIngestKey
		n := len(valueList)
		if n != 1 {
			siw.ErrorHandlerFunc(w, r, &TooManyValuesForParamError{ParamName: "X-Manifold-Ingest-Key", Count: n})
			return
		}

		err = runtime.BindStyledParameterWithOptions("simple", "X-Manifold-Ingest-Key", valueList[0], &XManifoldIngestKey, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: false})
		if err != nil {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "X-Manifold-Ingest-Key", Err: err})
			return
		}

		params.XManifoldIngestKey = &XManifoldIngestKey

	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostEventsBatch(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of http.ServeMux.
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	ServeHTTP(w http.ResponseWriter, r *http.Request)
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc("GET "+options.BaseURL+"/healthz", wrapper.GetHealthz)
	m.HandleFunc("GET "+options.BaseURL+"/readyz", wrapper.GetReadyz)
	m.HandleFunc("POST "+options.BaseURL+"/v1/events/batch", wrapper.PostEventsBatch)

	return m
}

type GetHealthzRequestObject struct {
}

type GetHealthzResponseObject interface {
	VisitGetHealthzResponse(w http.ResponseWriter) error
}

type GetHealthz200JSONResponse HealthResponse

func (response GetHealthz200JSONResponse) VisitGetHealthzResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetReadyzRequestObject struct {
}

type GetReadyzResponseObject interface {
	VisitGetReadyzResponse(w http.ResponseWriter) error
}

type GetReadyz200JSONResponse ReadyResponse

func (response GetReadyz200JSONResponse) VisitGetReadyzResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetReadyz503JSONResponse ReadyResponse

func (response GetReadyz503JSONResponse) VisitGetReadyzResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(503)

	return json.NewEncoder(w).Encode(response)
}

type PostEventsBatchRequestObject struct {
	Params PostEventsBatchParams
	Body   *PostEventsBatchJSONRequestBody
}

type PostEventsBatchResponseObject interface {
	VisitPostEventsBatchResponse(w http.ResponseWriter) error
}

type PostEventsBatch200JSONResponse APIResponse

func (response PostEventsBatch200JSONResponse) VisitPostEventsBatchResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type PostEventsBatch400JSONResponse APIResponse

func (response PostEventsBatch400JSONResponse) VisitPostEventsBatchResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type PostEventsBatch401JSONResponse APIResponse

func (response PostEventsBatch401JSONResponse) VisitPostEventsBatchResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)

	return json.NewEncoder(w).Encode(response)
}

type PostEventsBatch409JSONResponse APIResponse

func (response PostEventsBatch409JSONResponse) VisitPostEventsBatchResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(409)

	return json.NewEncoder(w).Encode(response)
}

type PostEventsBatch413JSONResponse APIResponse

func (response PostEventsBatch413JSONResponse) VisitPostEventsBatchResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(413)

	return json.NewEncoder(w).Encode(response)
}

type PostEventsBatch415JSONResponse APIResponse

func (response PostEventsBatch415JSONResponse) VisitPostEventsBatchResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(415)

	return json.NewEncoder(w).Encode(response)
}

type PostEventsBatch422JSONResponse APIResponse

func (response PostEventsBatch422JSONResponse) VisitPostEventsBatchResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(422)

	return json.NewEncoder(w).Encode(response)
}

type PostEventsBatch429JSONResponse APIResponse

func (response PostEventsBatch429JSONResponse) VisitPostEventsBatchResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(429)

	return json.NewEncoder(w).Encode(response)
}

type PostEventsBatch500JSONResponse APIResponse

func (response PostEventsBatch500JSONResponse) VisitPostEventsBatchResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type PostEventsBatch503JSONResponse APIResponse

func (response PostEventsBatch503JSONResponse) VisitPostEventsBatchResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(503)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Liveness probe
	// (GET /healthz)
	GetHealthz(ctx context.Context, request GetHealthzRequestObject) (GetHealthzResponseObject, error)
	// Readiness probe backed by storage ping
	// (GET /readyz)
	GetReadyz(ctx context.Context, request GetReadyzRequestObject) (GetReadyzResponseObject, error)
	// Ingest a batch of structured events
	// (POST /v1/events/batch)
	PostEventsBatch(ctx context.Context, request PostEventsBatchRequestObject) (PostEventsBatchResponseObject, error)
}

type StrictHandlerFunc = strictnethttp.StrictHTTPHandlerFunc
type StrictMiddlewareFunc = strictnethttp.StrictHTTPMiddlewareFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// GetHealthz operation middleware
func (sh *strictHandler) GetHealthz(w http.ResponseWriter, r *http.Request) {
	var request GetHealthzRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetHealthz(ctx, request.(GetHealthzRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetHealthz")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetHealthzResponseObject); ok {
		if err := validResponse.VisitGetHealthzResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetReadyz operation middleware
func (sh *strictHandler) GetReadyz(w http.ResponseWriter, r *http.Request) {
	var request GetReadyzRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetReadyz(ctx, request.(GetReadyzRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetReadyz")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetReadyzResponseObject); ok {
		if err := validResponse.VisitGetReadyzResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// PostEventsBatch operation middleware
func (sh *strictHandler) PostEventsBatch(w http.ResponseWriter, r *http.Request, params PostEventsBatchParams) {
	var request PostEventsBatchRequestObject

	request.Params = params

	var body PostEventsBatchJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.PostEventsBatch(ctx, request.(PostEventsBatchRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostEventsBatch")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(PostEventsBatchResponseObject); ok {
		if err := validResponse.VisitPostEventsBatchResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}